requirements.txt
streamlit
opencv-python
cryptography
Pillow